Perform relocations using inline asm.
Use inline asm to to the loads and stores needed by the relocation
logic, which happen outside the Rust memory model. This is conceptually
cleaner than using inline-asm to obscure the operands and result values
of the loads and stores, as it puts the actual memory access outside of
the scope of Rust.