Redesign `create_thread` to avoid dynamic allocation. (#94)
Change `create_thread` from taking a boxed closure to taking a function
pointer, as well as an array of pointers to be copied into the new thread
to pass to the function. This avoids the need to dynamically allocate for
a `Box`, and in particular it allows c-ward's `pthread_create` to be used
from within `malloc` implementations.

And, implement thread return values, which is simpler with the removal of
`Option<Box<dyn Any>>`.

These changes do create more work for users of the API, however it also
gives them more control.

Fixes #85.