Fix the location of the thread metadata on riscv64
On riscv64, the thread pointer doesn't point to the first ABI-exposed
field in the thread-local storage. Add an empty `thread_pointee` field
to represent the place the thread pointer points to, and fix the
address calculation code to point there.

And add a tls testcase that creates a `#[thread_local]` variable and
checks that it's initialized and can be accessed properly.