Move the relocate call from init_runtime to entry
Self-relocation is not actually initializing runtime state, but making
it possible to safely run any code at all. When using a dynamic linker
the dynamic linker needs to be relocated before it can do dynamic
linking, but some of the runtime intialization needs to happen after
dynamic linking when all static DSO's that use thread local storage are
known.