Couple of tls handling changes (#120)
* Pass module id to current_tls_addr

* Extract initialize_tls function to share between initialize_main and create

* Extract calculate_tls_size to share between initialize_main and create

The only functional change is that for spawned threads the padding
necessary to correctly align the TLS data is no longer added to the
stack itself and instead becomes unused. Most of the time TLS data will
be aligned no more than the stack alignment, so no memory should be
wasted most of the time.