Ensure self-relocation never panics
This is necessary as panicking would crash before relocations are
performed. I've manually verified that no panic calls are made in
release mode. When debug assertions are enabled there are still a couple
of panics inserted by the compiler, but these should be unreachable if
there are no bugs in the self-relocation code.