Various cleanups to the self-relocation code (#129)
* Use byte_add where possible

* Hardcode relocation entry sizes

Glibc only checks the DT_REL*ENT it in a debug assertion and musl
ignores them entirely, so we can be pretty sure that they will never
change. Hardcoding them makes the produced binaries a tiny bit smaller
and faster.

* Avoid all uses of with_exposed_provenance in relocate

By keeping around the original pointers rather than immediately
converting them to addresses. This is safe as none of these pointers,
nor their pointed to memory are affected by the relocations we are about
to perform.

* Ensure self-relocation never panics

This is necessary as panicking would crash before relocations are
performed. I've manually verified that no panic calls are made in
release mode. When debug assertions are enabled there are still a couple
of panics inserted by the compiler, but these should be unreachable if
there are no bugs in the self-relocation code.