Update to rustix 1.0. (#155)
* Update to rustix 1.0.

The main code change is that rustix 1.0 has better signals support.

Nightly rust removed the `#[start`] feature, so remove the "no-std"
example that used that. Other no-std options remain.

And it appears something in Rust changed how relocation of the
`_DYNAMIC` symbol works, so switch the relocation code to using
our assembler-defined version.

* Use `ehdr_addr()` instead of a compiler-managed static definition.